% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_carcdr.R
\name{qtls_apply_carcdr}
\alias{qtls_apply_carcdr}
\title{Apply to quosue tree
Applies a functions to car/cdr elements in quosure tree
on car/cdr's from a quosure}
\usage{
qtls_apply_carcdr(q, car_solo_apply = NULL, car_multi_apply = NULL,
  cdr_apply = NULL, context = NULL)
}
\arguments{
\item{parent}{}
}
\description{
Apply to quosue tree
Applies a functions to car/cdr elements in quosure tree
on car/cdr's from a quosure
}
