% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk.R
\name{qtls_graph_car_cdr_for_expression}
\alias{qtls_graph_car_cdr_for_expression}
\title{Title
build a table of parent child relationship based
on car/cdr's from a closure}
\usage{
qtls_graph_car_cdr_for_expression(expr)
}
\arguments{
\item{parent}{}
}
\description{
Title
build a table of parent child relationship based
on car/cdr's from a closure
}
