% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_info_function.R
\name{qtls_make_info_function}
\alias{qtls_make_info_function}
\title{Title}
\usage{
qtls_make_info_function(tbl, key_column, ...)
}
\arguments{
\item{tbl}{A table with a key column in it}

\item{key_column}{The column that serves as a key}

\item{...}{Columns used to make info string for row.
Values must be convertable to character.}
}
\value{
A function that returns a string
  with values (<column name>:<value>, )* for the row
  specified by the key passed into it
}
\description{
Title
}
