% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlineplot.R
\name{qtls_make_outline_plot}
\alias{qtls_make_outline_plot}
\title{Makes character based outline plot of tbl}
\usage{
qtls_make_outline_plot(tbl, key_column, ...)
}
\arguments{
\item{tbl}{a table with an id and parent column}

\item{key_column}{key column for table (typically id)}

\item{...}{columns to be included in output}
}
\description{
Makes character based outline plot of tbl
}
