% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_negate_om.R
\name{qtls_my_negate_om}
\alias{qtls_my_negate_om}
\title{Example using NSE for arithmetic expresion}
\usage{
qtls_my_negate_om(expr)
}
\arguments{
\item{expr}{an arithmetic expression}
}
\description{
Example using NSE for arithmetic expresion
}
