% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parent_child.R
\name{qtls_plot_model}
\alias{qtls_plot_model}
\title{Title
Builds a graph object using DiagrammeR for a tbl
that list obj id's and their parent.}
\usage{
qtls_plot_model(tbl, root_node = 1)
}
\arguments{
\item{tbl}{}

\item{root_node}{}
}
\description{
Table at least an id, parent, and atom column
}
\details{
The id column contains unique numeric identities for each row.

The parent column contains the id of the node of that row's parent.

The atom column contains text that tag's the nodes of the tree that
DiagrameR will use.
}
\examples{

}
