% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk.R
\name{qtls_walk_table}
\alias{qtls_walk_table}
\title{Title
Walks tree by following head/tail and produces tribble}
\usage{
qtls_walk_table(quosure, parent_id = 0, order = 1, context = NA)
}
\arguments{
\item{context}{}
}
\description{
Title
Walks tree by following head/tail and produces tribble
}
