% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rlang_model.R
\name{qtls_make_rlang_model}
\alias{qtls_make_rlang_model}
\title{Wrap the rlang object model for a language object in a table}
\usage{
qtls_make_rlang_model(e, parent = 0L, context = new.env(), order = 1L,
  path = vector(mode = "integer"))
}
\arguments{
\item{e}{a language object like an expr or quosure}

\item{parent}{parent of expression (used by recursion only)}

\item{context}{environment used to pass table being constructed
by reference}

\item{order}{position relative to siblings}

\item{path}{path in e to object}
}
\value{
A table where each row has, among other things a path
  to the object for that row
}
\description{
Wrap the rlang object model for a language object in a table
}
